+++ enterprise/cmd/frontend/internal/batches/webhooks/bitbucketcloud_test.go
@@ -15,11 +15,11 @@ import (
 
 	"github.com/sourcegraph/log/logtest"
 
-	bbcs "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/sources/bitbucketcloud"
-	bstore "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/store"
-	bt "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/testing"
-	btypes "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/types"
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	bbcs "github.com/sourcegraph/sourcegraph/internal/batches/sources/bitbucketcloud"
+	bstore "github.com/sourcegraph/sourcegraph/internal/batches/store"
+	bt "github.com/sourcegraph/sourcegraph/internal/batches/testing"
+	btypes "github.com/sourcegraph/sourcegraph/internal/batches/types"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"